[arm] fix segfault of multithreaded Qt programs in thumb2 mode
This fixes a segfault when using QMutex in a multithreaded program.
The problem only shows up for ARM with Thumb2 and CONFIG_USER_ONLY.

The accounting of instructions in an IT (if-then) block went
off by one when a STREX exception was generated. This caused
the instruction after the block to be mistakenly compiled as
part of the block, and in case of a conditional branch this
meant generating a branch to a label that is never defined.

Signed-off-by: Richard Braakman <richard.braakman@jollamobile.com>